Dev mode hook
